# perf(request): optimize if headers are given