--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -120,6 +120,19 @@ function appendHeader (headers, name, value) {
   //    privileged no-CORS request headers from headers
 }
 
+/**
+ * @param {HeadersList} A copy destination
+ * @param {HeadersList} B copy source
+ */
+function headersListCopy (A, B) {
+  if (B[kHeadersMap].size !== 0) {
+    for (const [key, { name, value }] of B[kHeadersMap]) {
+      A[kHeadersMap].set(key, { name, value })
+    }
+    A.cookies = B.cookies === null ? null : [...B.cookies]
+  }
+}
+
 class HeadersList {
   /** @type {[string, string][]|null} */
   cookies = null
@@ -222,6 +235,10 @@ class HeadersList {
     for (const [name, { value }] of this[kHeadersMap]) {
       yield [name, value]
     }
+  }
+
+  get mapSize () {
+    return this[kHeadersMap].size
   }
 
   get entries () {
@@ -581,6 +598,7 @@ webidl.converters.HeadersInit = function (V) {
 
 module.exports = {
   fill,
+  headersListCopy,
   Headers,
   HeadersList
 }