--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -126,8 +126,8 @@ function appendHeader (headers, name, value) {
  */
 function headersListCopy (A, B) {
   if (B[kHeadersMap].size !== 0) {
-    for (const [key, { name, value }] of B[kHeadersMap]) {
-      A[kHeadersMap].set(key, { name, value })
+    for (const entry of B[kHeadersMap]) {
+      A[kHeadersMap].set(entry[0], entry[1])
     }
     A.cookies = B.cookies === null ? null : [...B.cookies]
   }