--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -129,8 +129,8 @@ function headersListCopy (A, B) {
     for (const entry of B[kHeadersMap]) {
       A[kHeadersMap].set(entry[0], entry[1])
     }
-    A.cookies = B.cookies === null ? null : [...B.cookies]
   }
+  A.cookies = B.cookies === null ? null : [...B.cookies]
 }
 
 class HeadersList {