--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -3,7 +3,7 @@
 'use strict'
 
 const { extractBody, mixinBody, cloneBody } = require('./body')
-const { Headers, fill: fillHeaders, HeadersList } = require('./headers')
+const { Headers, fill: fillHeaders, HeadersList, headersListCopy } = require('./headers')
 const { FinalizationRegistry } = require('../compat/dispatcher-weakref')()
 const util = require('../core/util')
 const {
@@ -415,25 +415,23 @@ class Request {
     }
 
     // 32. If init is not empty, then:
-    if (Object.keys(init).length !== 0) {
+    // Note: Avoid call `Object.keys` if init["headers"] is not undefined.
+    if (init.headers !== undefined || Object.keys(init).length !== 0) {
       // 1. Let headers be a copy of this’s headers and its associated header
       // list.
-      let headers = new Headers(this[kHeaders])
-
       // 2. If init["headers"] exists, then set headers to init["headers"].
-      if (init.headers !== undefined) {
-        headers = init.headers
-      }
+      // Note: Avoid call `HeadersList` if init["headers"] is not undefined
+      const headers = init.headers !== undefined ? init.headers : new HeadersList(request.headersList)
 
       // 3. Empty this’s headers’s header list.
-      this[kHeaders][kHeadersList].clear()
+      // Note: Avoid call `HeadersList#clear` if already empty
+      if (request.headersList.mapSize !== 0) request.headersList.clear()
 
       // 4. If headers is a Headers object, then for each header in its header
       // list, append header’s name/header’s value to this’s headers.
-      if (headers.constructor.name === 'Headers') {
-        for (const [key, val] of headers) {
-          this[kHeaders].append(key, val)
-        }
+      // Note: If init["headers"] is undefined, then it is `HeadersList`.
+      if (init.headers === undefined) {
+        headersListCopy(request.headersList, headers)
       } else {
         // 5. Otherwise, fill this’s headers with headers.
         fillHeaders(this[kHeaders], headers)