--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -183,8 +183,10 @@ class Request {
       urlList: [...request.urlList]
     })
 
+    const initHasKey = Object.keys(init).length !== 0
+
     // 13. If init is not empty, then:
-    if (Object.keys(init).length > 0) {
+    if (initHasKey) {
       // 1. If request’s mode is "navigate", then set it to "same-origin".
       if (request.mode === 'navigate') {
         request.mode = 'same-origin'
@@ -415,23 +417,23 @@ class Request {
     }
 
     // 32. If init is not empty, then:
-    // Note: Avoid call `Object.keys` if init["headers"] is not undefined.
-    if (init.headers !== undefined || Object.keys(init).length !== 0) {
+    if (initHasKey) {
+      const headersList = this[kHeaders][kHeadersList]
       // 1. Let headers be a copy of this’s headers and its associated header
       // list.
       // 2. If init["headers"] exists, then set headers to init["headers"].
       // Note: Avoid call `HeadersList` if init["headers"] is not undefined
-      const headers = init.headers !== undefined ? init.headers : new HeadersList(request.headersList)
+      const headers = init.headers !== undefined ? init.headers : new HeadersList(headersList)
 
       // 3. Empty this’s headers’s header list.
       // Note: Avoid call `HeadersList#clear` if already empty
-      if (request.headersList.mapSize !== 0) request.headersList.clear()
+      if (headersList.mapSize !== 0) headersList.clear()
 
       // 4. If headers is a Headers object, then for each header in its header
       // list, append header’s name/header’s value to this’s headers.
       // Note: If init["headers"] is undefined, then it is `HeadersList`.
       if (init.headers === undefined) {
-        headersListCopy(request.headersList, headers)
+        headersListCopy(headersList, headers)
       } else {
         // 5. Otherwise, fill this’s headers with headers.
         fillHeaders(this[kHeaders], headers)