--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -3,7 +3,7 @@
 'use strict'
 
 const { extractBody, mixinBody, cloneBody } = require('./body')
-const { Headers, fill: fillHeaders, HeadersList, headersListCopy } = require('./headers')
+const { Headers, fill: fillHeaders, HeadersList } = require('./headers')
 const { FinalizationRegistry } = require('../compat/dispatcher-weakref')()
 const util = require('../core/util')
 const {
@@ -418,22 +418,28 @@ class Request {
 
     // 32. If init is not empty, then:
     if (initHasKey) {
+      /** @type {HeadersList} */
       const headersList = this[kHeaders][kHeadersList]
       // 1. Let headers be a copy of this’s headers and its associated header
       // list.
       // 2. If init["headers"] exists, then set headers to init["headers"].
-      // Note: Avoid call `HeadersList` if init["headers"] is not undefined
-      const headers = init.headers !== undefined ? init.headers : new HeadersList(headersList)
+      let headers
+      if (init.headers !== undefined) {
+        headers = init.headers
+      } else {
+        init.headers = new HeadersList(headersList)
+      }
 
       // 3. Empty this’s headers’s header list.
-      // Note: Avoid call `HeadersList#clear` if already empty
-      if (headersList.mapSize !== 0) headersList.clear()
+      headersList.clear()
 
       // 4. If headers is a Headers object, then for each header in its header
       // list, append header’s name/header’s value to this’s headers.
-      // Note: If init["headers"] is undefined, then it is `HeadersList`.
-      if (init.headers === undefined) {
-        headersListCopy(headersList, headers)
+      if (headers instanceof HeadersList) {
+        for (const entry of headers) {
+          headersList.set(entry[0], entry[1])
+        }
+        headersList.cookies = headers.cookies
       } else {
         // 5. Otherwise, fill this’s headers with headers.
         fillHeaders(this[kHeaders], headers)