--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -422,8 +422,8 @@ class Request {
       const headersList = this[kHeaders][kHeadersList]
       // 1. Let headers be a copy of this’s headers and its associated header
       // list.
-      // 2. If init["headers"] exists, then set headers to init["headers"].
       let headers
+      // 2. If init["headers"] exists, then set headers to init["headers"].
       if (init.headers !== undefined) {
         headers = init.headers
       } else {
@@ -436,8 +436,8 @@ class Request {
       // 4. If headers is a Headers object, then for each header in its header
       // list, append header’s name/header’s value to this’s headers.
       if (headers instanceof HeadersList) {
-        for (const entry of headers) {
-          headersList.set(entry[0], entry[1])
+        for (const [key, value] of headers) {
+          headersList.set(key, value)
         }
         headersList.cookies = headers.cookies
       } else {