--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -423,6 +423,7 @@ class Request {
       // 1. Let headers be a copy of this’s headers and its associated header
       // list.
       let headers
+
       // 2. If init["headers"] exists, then set headers to init["headers"].
       if (init.headers !== undefined) {
         headers = init.headers