--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -437,8 +437,8 @@ class Request {
       // 4. If headers is a Headers object, then for each header in its header
       // list, append header’s name/header’s value to this’s headers.
       if (headers instanceof HeadersList) {
-        for (const [key, value] of headers) {
-          headersList.set(key, value)
+        for (const [key, val] of headers) {
+          headersList.set(key, val)
         }
         headersList.cookies = headers.cookies
       } else {