--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -428,7 +428,7 @@ class Request {
       if (init.headers !== undefined) {
         headers = init.headers
       } else {
-        init.headers = new HeadersList(headersList)
+        headers = new HeadersList(headersList)
       }
 
       // 3. Empty this’s headers’s header list.