--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -440,6 +440,7 @@ class Request {
         for (const [key, val] of headers) {
           headersList.set(key, val)
         }
+        // Note: Copy the `set-cooke` meta-data.
         headersList.cookies = headers.cookies
       } else {
         // 5. Otherwise, fill this’s headers with headers.