--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -438,7 +438,7 @@ class Request {
       // list, append header’s name/header’s value to this’s headers.
       if (headers instanceof HeadersList) {
         for (const [key, val] of headers) {
-          headersList.set(key, val)
+          headersList.append(key, val)
         }
         // Note: Copy the `set-cooke` meta-data.
         headersList.cookies = headers.cookies