--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -422,14 +422,8 @@ class Request {
       const headersList = this[kHeaders][kHeadersList]
       // 1. Let headers be a copy of this’s headers and its associated header
       // list.
-      let headers
-
       // 2. If init["headers"] exists, then set headers to init["headers"].
-      if (init.headers !== undefined) {
-        headers = init.headers
-      } else {
-        headers = new HeadersList(headersList)
-      }
+      const headers = init.headers !== undefined ? init.headers : new HeadersList(headersList)
 
       // 3. Empty this’s headers’s header list.
       headersList.clear()