<!--
Before submitting a Pull Request, please read our contribution guidelines, which
can be found at CONTRIBUTING.md in the repository root.

For code changes:
1. Include tests for any bug fixes or new features.
2. Update documentation if relevant.
3. Ensure that tests and linting pass.

You will also need to ensure that your contribution complies with the
Developer's Certificate of Origin, outlined in CONTRIBUTING.md
-->

## Rationale

<!-- Briefly explain the purpose of this pull request, if not already
justifiable with the above section. If it is, you may omit this section. -->

For performance improvement:
1. Do not call `Object.keys` more than once.
2. When copying the `headersList`, use `HeadersList` instead of `Headers`.
3. Avoid call `HeadersList` if init["headers"] is not undefined
4. When looping through headers, use `HeadersList` instead of `Headers`, which eliminates the sorting process and is faster.

## Benchmark

```js
// bench.js
const { Request } = require(".");
const benchmark = require("benchmark");

const suite = new benchmark.Suite();

const headers = {
  "Content-Type": "application/json",
  "X-Content-Type": "application/json",
  Date: "Wed, 01 Nov 2023 00:00:00 GMT",
  "X-NodeJS-Version": "v21.1.0",
  "Powered-By": "NodeJS",
  "Content-Encoding": "gzip",
  "Set-Cookie": "__Secure-ID=123; Secure; Domain=example.com",
  "Content-Length": "150",
  Vary: "Accept-Encoding, Accept, X-Requested-With",
};

const headersEntries = Object.entries(headers);
suite.add("new Request() - record", () => {
  new Request("https://example.com/post", {
    headers: headers,
  });
});

suite.add("new Request() - record without constructor", () => {
  const request = new Request("https://example.com/post", {});
  const h = request.headers
  const keys = Object.keys(headers);
  for (let i = 0; i < keys.length; ++i) {
    h.append(keys[i], headers[keys[i]]);
  } 
})

suite.add("new Request() - entires", () => {
  new Request("https://example.com/post", {
    headers: headersEntries,
  });
});

suite.add("new Request() - entires without constructor", () => {
  const request = new Request("https://example.com/post", {});
  const h = request.headers
  for (let i = 0; i < headersEntries.length; ++i) {
    h.append(headersEntries[i][0], headersEntries[i][1]);
  }
});

suite.on("cycle", function (event) {
  console.log(String(event.target));
});

// run async
suite.run({ async: true });
```

- *v5.27.2*
```
new Request() - record x 34,814 ops/sec ±3.33% (75 runs sampled)
new Request() - record without constructor x 64,510 ops/sec ±2.22% (85 runs sampled)
new Request() - entires x 34,316 ops/sec ±3.31% (77 runs sampled)
new Request() - entires without constructor x 61,662 ops/sec ±3.00% (79 runs sampled)
```

- *main*
```
new Request() - record x 41,193 ops/sec ±5.16% (75 runs sampled)
new Request() - record without constructor x 65,217 ops/sec ±2.69% (80 runs sampled)
new Request() - entires x 38,771 ops/sec ±2.78% (78 runs sampled)
new Request() - entires without constructor x 68,280 ops/sec ±2.86% (83 runs sampled)
```

- *this patch*
```
new Request() - record x 68,534 ops/sec ±2.08% (80 runs sampled)
new Request() - record without constructor x 69,161 ops/sec ±2.31% (83 runs sampled)
new Request() - entires x 65,713 ops/sec ±2.13% (84 runs sampled)
new Request() - entires without constructor x 67,820 ops/sec ±2.57% (79 runs sampled)
```

## Status

<!-- KEY: S = Skipped, x = complete -->


- [ ] I have read and agreed to the [Developer's Certificate of Origin][cert]
- [ ] Tested
- [ ] Benchmarked (**optional**)
- [ ] Documented
- [ ] Review ready
- [ ] In review
- [ ] Merge ready

[cert]: https://github.com/nodejs/undici/blob/main/CONTRIBUTING.md