@mcollina 2023-11-24
Does this pass the same number of wpts?