@tsctx 2023-11-24
@mcollina 
See here
- *main*
```
[fetch]: Completed: 1453, failed: 165, success: 1288, expected failures: 164, unexpected failures: 1, skipped: 71
[FileAPI]: Completed: 406, failed: 16, success: 390, expected failures: 16, unexpected failures: 0, skipped: 3
[mimesniff]: Completed: 1898, failed: 872, success: 1026, expected failures: 872, unexpected failures: 0, skipped: 0
[xhr/formdata]: Completed: 39, failed: 0, success: 39, expected failures: 0, unexpected failures: 0, skipped: 0
[websockets]: Completed: 500, failed: 18, success: 482, expected failures: 18, unexpected failures: 0, skipped: 10
```
- *this patch*
```
[fetch]: Completed: 1453, failed: 165, success: 1288, expected failures: 164, unexpected failures: 1, skipped: 71
[FileAPI]: Completed: 406, failed: 16, success: 390, expected failures: 16, unexpected failures: 0, skipped: 3
[mimesniff]: Completed: 1898, failed: 872, success: 1026, expected failures: 872, unexpected failures: 0, skipped: 0
[xhr/formdata]: Completed: 39, failed: 0, success: 39, expected failures: 0, unexpected failures: 0, skipped: 0
[websockets]: Completed: 500, failed: 18, success: 482, expected failures: 18, unexpected failures: 0, skipped: 10
```