@KhafraDev 2023-11-24
Spec breakage isn't the concern here, it's not following the spec directly that worries me. If for example those lines are changed in the spec, how do we upstream the changes here? It makes maintaining it a nightmare. If the change is completely compatible with the spec, I don't see a reason that the changes couldn't be made in the fetch spec itself, and then made here.