@KhafraDev 2023-11-24
I have no problem with something like setting the value of `Object.keys(init).length` to a variable, but right now the comments from the spec are no longer what undici is doing. I promise that if we don't follow the spec directly, it'll lead to an unmaintainable mess (ie. look at basically any spec node core implements).