@mcollina 2023-11-25
@KhafraDev these APIs are called by a _lot_, making a small improvement noticeable perf-wise.